+++ src/Proxy.hs
@@ -58,7 +58,6 @@ import Data.Word
 import qualified Database.Esqueleto.Internal.Internal as EInternal
 import qualified Database.Esqueleto.Legacy as E
 import qualified Database.Esqueleto.PostgreSQL.JSON as EJ
-import Debug.Trace (traceM) -- temporarly
 import Lens.Micro.Platform hiding ((.=))
 import Network.GRPC.HTTP2.Types (GRPCStatusCode (..))
 import Network.HTTP.Types (Status, badGateway502, badRequest400, internalServerError500, notFound404, serviceUnavailable503)
@@ -87,6 +86,7 @@ import qualified Concordium.Wasm as Wasm
 import Paths_wallet_proxy (version)
 
 import Concordium.Client.GRPC2
+import Concordium.Client.Runner (normalizeTokenAmountOrDie)
 import Concordium.Client.Runner.Helper
 import Concordium.Client.Types.Transaction (
     accountDecryptEnergyCost,
@@ -109,7 +109,7 @@ import Concordium.Client.Types.Transaction (
     transferWithSchedulePayloadSize,
     updateDelegationPayloadSize,
  )
-import Concordium.Client.Utils (tokenAmountFromString)
+import Concordium.Client.Utils (preTokenAmountFromStringInform)
 import Concordium.Common.Version
 import Concordium.Crypto.ByteStringHelpers (ShortByteStringHex (..))
 import Concordium.Crypto.SHA256 (Hash)
@@ -833,11 +833,29 @@ getAccountEncryptionKeyR addrText = do
                     <> Text.pack (show encryptionKey)
             sendResponse (object ["accountEncryptionKey" .= encryptionKey])
 
+-- The function looks up the `memoSize` query parameter and computes the additional size of the transaction based on the memo.
+-- This `memoSize` paramater can only be applied to transfer and encrypted transfer transaction types and is
+-- only supported if the node is running protocol version 2 or higher.
+getMemoPayloadSize :: ProtocolVersion -> Handler PayloadSize
+getMemoPayloadSize pv = do
+    lookupGetParam "memoSize" >>= \case
+        Nothing -> return 0
+        Just memoText ->
+            case readMaybe (Text.unpack memoText) :: Maybe Word32 of
+                Nothing -> respond400Error (EMParseError "Could not parse `memoSize` value.") RequestInvalid
+                -- NB: In protocol version 1 the memo is not supported. This
+                -- implementation assumes that the transaction memo will be
+                -- supported in all future versions of the node.
+                -- The memo is charged for purely on its size. The "2 +" is there because the memo
+                -- is serialized by prepending 2 bytes for its length.
+                Just msize
+                    | pv /= P1 -> return $ fromIntegral (2 + msize)
+                    | otherwise -> respond404Error EMActionNotCurrentlySupported
+
 -- | Get the cost of a transaction, based on its type. The following query parameters are supported
 --  - "type", the type of the transaction. This is mandatory.
 --  - "numSignatures", the number of signatures on the transaction, defaults to 1 if not present.
---  - "memoSize", the size of the transfer memo. Only supported if the node is running protocol version 2 or higher, and
---    only applies when `type` is either `simpleTransfer`, `encryptedTransfer`, or `simplePltTransfer`.
+--  - additional transaction type specific query parameters
 getTransactionCostR :: Handler TypedContent
 getTransactionCostR = withExchangeRate $ \(rate, pv) -> do
     numSignatures <- fromMaybe "1" <$> lookupGetParam "numSignatures"
@@ -847,22 +865,7 @@ getTransactionCostR = withExchangeRate $ \(rate, pv) -> do
   where
     handleTransactionCost pv rate numSignatures = do
         transactionType <- lookupGetParam "type"
-        -- compute the additional size of the transaction based on the memo
-        -- this only applies to transfer and encrypted transfer transaction types.
-        memoPayloadSize <- do
-            lookupGetParam "memoSize" >>= \case
-                Nothing -> return 0
-                Just memoText ->
-                    case readMaybe (Text.unpack memoText) :: Maybe Word32 of
-                        Nothing -> respond400Error (EMParseError "Could not parse `memoSize` value.") RequestInvalid
-                        -- NB: In protocol version 1 the memo is not supported. This
-                        -- implementation assumes that the transaction memo will be
-                        -- supported in all future versions of the node.
-                        -- The memo is charged for purely on its size. The "2 +" is there because the memo
-                        -- is serialized by prepending 2 bytes for its length.
-                        Just msize
-                            | pv /= P1 -> return $ fromIntegral (2 + msize)
-                            | otherwise -> respond404Error EMActionNotCurrentlySupported
+
         let costResponse energyCost =
                 sendResponse $
                     object
@@ -879,23 +882,24 @@ getTransactionCostR = withExchangeRate $ \(rate, pv) -> do
         case transactionType of
             Nothing -> respond400Error (EMParseError "Missing `type` value.") RequestInvalid
             Just tty -> case Text.unpack tty of
-                "simpleTransfer" ->
+                "simpleTransfer" -> do
+                    memoPayloadSize <- getMemoPayloadSize pv
                     costResponse $ simpleTransferEnergyCost (simpleTransferPayloadSize + memoPayloadSize) numSignatures
                 "simplePltTransfer" -> do
                     amount <-
-                        lookupGetParam "amount" >>= \case
-                            Nothing -> respond400Error (EMParseError "Missing `amount` value.") RequestInvalid
+                        lookupGetParam "tokenAmount" >>= \case
+                            Nothing -> respond400Error (EMParseError "Missing `tokenAmount` value.") RequestInvalid
                             Just a -> case readMaybe $ Text.unpack a of
-                                Nothing -> respond400Error (EMParseError "Could not parse `amount` value.") RequestInvalid
+                                Nothing -> respond400Error (EMParseError "Could not parse `tokenAmount` value.") RequestInvalid
                                 Just b -> return b
-
                     maybeTextMemo <- do
                         lookupGetParam "textMemo" >>= \case
                             Nothing -> return Nothing
                             Just a -> case readMaybe $ Text.unpack a of
-                                Nothing -> respond400Error (EMParseError "Could not parse `memo` value.") RequestInvalid
+                                Nothing -> respond400Error (EMParseError "Could not parse `textMemo` value.") RequestInvalid
                                 -- The wallets currently submit the `TaggableMemo` type using the `UntaggedMemo` variant.
                                 Just b -> return $ Just (CBOR.UntaggedMemo (Memo (BSS.toShort (Text.encodeUtf8 b))))
+                    let maybeMemoSize = fmap (BS.length . CBOR.taggableMemoToBytes) maybeTextMemo
 
                     tokenId <-
                         lookupGetParam "tokenId" >>= \case
@@ -904,16 +908,22 @@ getTransactionCostR = withExchangeRate $ \(rate, pv) -> do
                                 Nothing -> respond400Error (EMParseError "Could not parse `tokenId` value.") RequestInvalid
                                 Just b -> return b
 
-                    let tokenAmount = tokenAmountFromString amount
-                    tokenAmountSize <- case tokenAmount of
-                        Just value -> pure (BS.length $ CBOR.tokenAmountToBytes value)
-                        Nothing -> respond400Error (EMParseError "Invalid `tokenAmount` value.") RequestInvalid
+                    let tokenAmount = preTokenAmountFromStringInform amount
+                    tAmount <- case tokenAmount of
+                        Left err ->
+                            respond400Error (EMParseError ("Couldn't parse `tokenAmount`: " <> err)) RequestInvalid
+                        Right b -> pure b
+                    tokenInfo <- runGRPC (getTokenInfoFromText tokenId LastFinal) pure
+                    normAmount <- normalizeTokenAmountOrDie tokenInfo tAmount
+                    let tokenAmountSize = BS.length $ CBOR.tokenAmountToBytes normAmount
 
-                    let maybeMemoSize = fmap (BS.length . CBOR.taggableMemoToBytes) maybeTextMemo
-
+                    -- ============================
+                    -- BEGIN ADDITIONAL COMMENTS
+                    -- ============================
+                    --
                     -- The wallets currently submit the `tokenHolder` value without `coinInfo`.
                     -- If the wallets submit the value with `coinInfo` one day, use the following calculation below instead:
-
+                    --
                     -- CBOR byte sequence of `tokenHolder` as follows (with coinInfo):
                     -- - d99d73 a2: A tagged (40307) item containing a map with 2 key-value pairs
                     --  - 01 d99d71 a1: Key 1 => d99d71: A tagged (40305) item containing a map with 1 key-value pair:
@@ -925,6 +935,10 @@ getTransactionCostR = withExchangeRate $ \(rate, pv) -> do
                     --       01 190397 (4 bytes)
                     --     03 5820 1515151515151515151515151515151515151515151515151515151515151515 (35 bytes)
                     -- let tokenHolderSize = 48
+                    --
+                    -- ==========================
+                    -- END ADDITIONAL COMMENTS
+                    -- ==========================
 
                     -- The wallets currently submit the `tokenHolder` value without `coinInfo`.
 
@@ -936,7 +950,7 @@ getTransactionCostR = withExchangeRate $ \(rate, pv) -> do
                     --     03 5820 1515151515151515151515151515151515151515151515151515151515151515 (35 bytes)
                     let tokenHolderSize = 39
 
-                    -- Payload of a tokenUpdate transaction encoding exactly one plt transfer operation (without memo).
+                    -- Payload of a tokenUpdate transaction encoding exactly one plt transfer operation.
                     let baseSize =
                             -- 1 byte for Map header
                             1
@@ -975,7 +989,7 @@ getTransactionCostR = withExchangeRate $ \(rate, pv) -> do
 
                     let tokenIdSize =
                             -- 1 byte for length
-                            1 + BS.length tokenId
+                            1 + BS.length (Text.encodeUtf8 tokenId)
 
                     let tokenUpdateTransactionSize =
                             -- 1 byte (tag) to specify the `TransactionType`.
@@ -983,9 +997,9 @@ getTransactionCostR = withExchangeRate $ \(rate, pv) -> do
                                 + tokenIdSize
                                 + tokenParameterSize
 
-                    traceM $ "tokenUpdateSize: " ++ show tokenUpdateTransactionSize
                     costResponse $ tokenUpdateTransactionEnergyCost (fromIntegral tokenUpdateTransactionSize) Cost.tokenTransferCost numSignatures
-                "encryptedTransfer" ->
+                "encryptedTransfer" -> do
+                    memoPayloadSize <- getMemoPayloadSize pv
                     costResponse $ encryptedTransferEnergyCost (encryptedTransferPayloadSize + memoPayloadSize) numSignatures
                 "transferToSecret" ->
                     costResponse $ accountEncryptEnergyCost accountEncryptPayloadSize numSignatures